moved the scene to graphics
